ZEPPELIN-501: refactoring, extract SearchService interface + Lucen-based impl
